requirements.txt
Django==4.1.2
django-import-export==3.0.1
